+++ fltk/src/macros/table.rs
@@ -72,6 +72,15 @@ macro_rules! impl_table_ext {
                     }
                 }
 
+                fn try_visible_cells(&self) -> Option<(i32, i32, i32, i32)> {
+                    let (a, b, c, d) = self.visible_cells();
+                    if a == -1 || b == -1 || c == -1 || d == -1 {
+                        None
+                    } else {
+                        Some((a, b, c, d))
+                    }
+                }
+
                 fn is_interactive_resize(&self) -> bool {
                     unsafe {
                         assert!(!self.was_deleted());
@@ -328,6 +337,17 @@ macro_rules! impl_table_ext {
                     }
                 }
 
+                fn try_get_selection(&self) -> Option<(i32, i32, i32, i32)> {
+                    let (a, b, c, d) = self.get_selection();
+                    if a < 0 && b < 0 && c >= 0 && d >= 0 {
+                        Some((0, 0, c, d))
+                    } else if a >= 0 && b >=0 && c >=0 && d >= 0 {
+                        Some((a, b, c, d))
+                    } else {
+                        None
+                    }
+                }
+
                 fn set_selection(&mut self, row_top: i32, col_left: i32, row_bot: i32, col_right: i32) {
                     unsafe {
                         assert!(!self.was_deleted());
@@ -409,6 +429,7 @@ macro_rules! impl_table_ext {
                     cb: F,
                 ) {
                     assert!(!self.was_deleted());
+                    assert!(self.is_derived);
                     pub type CustomDrawCellCallback = Option<
                         unsafe extern "C" fn(
                             wid: *mut Fl_Widget,
@@ -517,6 +538,7 @@ macro_rules! impl_table_ext {
 
                 fn scrollbar(&self) -> $crate::valuator::Scrollbar {
                     assert!(!self.was_deleted());
+                    assert!(self.is_derived);
                     unsafe {
                         let ptr = [<$flname _scrollbar>](self.inner);
                         assert!(!ptr.is_null());
@@ -528,12 +550,35 @@ macro_rules! impl_table_ext {
 
                 fn hscrollbar(&self) -> $crate::valuator::Scrollbar {
                     assert!(!self.was_deleted());
+                    assert!(self.is_derived);
                     unsafe {
                         let ptr = [<$flname _hscrollbar>](self.inner);
                         assert!(!ptr.is_null());
                         $crate::valuator::Scrollbar::from_widget_ptr(
                             ptr as *mut fltk_sys::widget::Fl_Widget,
                         )
+                    }
+                }
+
+                fn find_cell(
+                    &self,
+                    ctx: crate::table::TableContext,
+                    row: i32,
+                    col: i32,
+                ) -> Option<(i32, i32, i32, i32)> {
+                    assert!(!self.was_deleted());
+                    assert!(self.is_derived);
+                    let mut x = 0;
+                    let mut y = 0;
+                    let mut w = 0;
+                    let mut h = 0;
+                    unsafe {
+                        let ret = [<$flname _find_cell>](self.inner, ctx as i32, row, col, &mut x, &mut y, &mut w, &mut h);
+                        if ret == 0 {
+                            Some((x, y, w, h))
+                        } else {
+                            None
+                        }
                     }
                 }
             }